+++ index.html
@@ -67,6 +67,14 @@
 
     <title>PyCamp España</title>
 
+    <!--
+      JSON-LD Event schema for the active edition.
+      Re-enable (and update) when the next edition is confirmed:
+        - "name", "description", "startDate", "endDate"
+        - "image" URL path (e.g. /2027/...)
+        - "offers.name", "offers.url", "offers.validFrom"
+        - "url" if the event page slug changes
+
     <script type="application/ld+json">
       {
         "@context": "https://schema.org",
@@ -109,6 +117,7 @@
         "url": "https://pycamp.es/event"
       }
     </script>
+    -->
   </head>
 
   <body>
