+++ src/pages/Home/Home.css
@@ -125,182 +125,6 @@
   line-height: 1.8;
 }
 
-/* Newsletter Section */
-.newsletter {
-  background: linear-gradient(
-    180deg,
-    rgba(246, 244, 238, 1) 0%,
-    rgba(250, 248, 245, 1) 36%
-  );
-  padding: var(--spacing-xl) var(--spacing-md);
-}
-
-.newsletter-container {
-  max-width: 850px;
-  margin: 0 auto;
-  padding: var(--spacing-md) var(--spacing-lg);
-  text-align: center;
-  background: var(--surface-subtle);
-  border: 2px solid var(--surface-muted);
-  border-radius: var(--radius-md);
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacing-sm);
-  box-shadow: 0 1px 20px rgba(238, 231, 221, 1);
-}
-
-.newsletter-icon {
-  width: 58px;
-  height: 58px;
-  background: rgba(244, 123, 37, 0.18);
-  border-radius: var(--radius-sm);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 auto var(--spacing-md);
-  color: var(--accent-campfire);
-}
-
-.newsletter-icon img {
-  width: 34px;
-  height: 34px;
-  object-fit: contain;
-}
-
-.newsletter-title {
-  font-size: 32px;
-  margin-bottom: var(--spacing-sm);
-}
-
-.newsletter-description {
-  color: var(--text-secondary);
-  font-size: 16px;
-  margin-bottom: var(--spacing-sm);
-  line-height: 1.8;
-}
-
-.newsletter-form {
-  max-width: 850px;
-  margin: 0 auto;
-  padding: var(--spacing-sm);
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacing-sm);
-}
-
-.newsletter-form-row {
-  display: flex;
-  gap: var(--spacing-sm);
-  flex-wrap: wrap;
-}
-
-.newsletter-input {
-  flex: 1;
-  min-width: 220px;
-  border: 2px solid var(--surface-accent);
-  border-radius: var(--radius-md);
-  padding: 13px 28px;
-  font-size: 20px;
-  background: rgba(250, 248, 245, 0.2);
-  color: var(--text-primary);
-}
-
-.newsletter-input::placeholder {
-  color: var(--ui-border);
-}
-
-.newsletter-input:focus {
-  outline: none;
-  border-color: var(--brand-green-dark);
-}
-
-.newsletter-consent {
-  display: flex;
-  align-items: flex-start;
-  gap: var(--spacing-xs);
-  font-size: 0.85rem;
-  line-height: 1.4;
-  color: var(--text-muted);
-  cursor: pointer;
-  margin-top: var(--spacing-xs);
-}
-
-.newsletter-consent input[type="checkbox"] {
-  position: absolute;
-  opacity: 0;
-  pointer-events: none;
-}
-
-.newsletter-consent span {
-  position: relative;
-  padding-left: 28px;
-}
-
-.newsletter-consent span::before {
-  content: "";
-  position: absolute;
-  left: 0;
-  top: 0.15em;
-  width: 18px;
-  height: 18px;
-  border-radius: var(--radius-sm);
-  background: var(--surface-background);
-  border: 2px solid var(--ui-border);
-}
-
-.newsletter-consent input:checked + span::before {
-  background: var(--accent-forest);
-  border-color: var(--brand-green-medium);
-}
-
-.newsletter-consent span::after {
-  content: "✓";
-  position: absolute;
-  left: 2px;
-  top: 0.25em;
-  width: 18px;
-  height: 18px;
-  font-size: 14px;
-  font-weight: 700;
-  line-height: 18px;
-  text-align: center;
-  color: var(--brand-green-dark);
-  opacity: 0;
-  transform: scale(0.6);
-  transition:
-    opacity 0.15s ease,
-    transform 0.15s ease;
-}
-
-.newsletter-consent input:checked + span::after {
-  opacity: 1;
-  transform: scale(1);
-}
-
-.newsletter-success {
-  padding: var(--spacing-md);
-  background: var(--accent-forest);
-  border-radius: var(--radius-md);
-  text-align: center;
-}
-
-.newsletter-success h3 {
-  font-family: "Yeseva One", serif;
-  color: var(--brand-green-dark);
-  margin-bottom: var(--spacing-xs);
-}
-
-.newsletter-success p {
-  font-size: 0.95rem;
-  color: var(--text-primary);
-}
-
-.newsletter-legal {
-  margin-top: var(--spacing-xs);
-  font-size: 0.85rem;
-  color: var(--text-muted);
-}
-
 /* Sponsors Section */
 .sponsors {
   background: var(--surface-background);
@@ -340,7 +164,6 @@
 
   .hero-container,
   .about-container,
-  .newsletter-container,
   .sponsors-container {
     padding: var(--spacing-md) var(--spacing-md);
   }
@@ -349,7 +172,6 @@
 @media (max-width: 768px) {
   .hero-container,
   .about-container,
-  .newsletter-container,
   .sponsors-container {
     padding: var(--spacing-sm) var(--spacing-sm);
   }
@@ -364,15 +186,6 @@
 
   .about-title {
     font-size: 32px;
-  }
-
-  .newsletter-form {
-    flex-direction: column;
-  }
-  .newsletter-form-row {
-    display: flex;
-    align-items: center;
-    justify-content: center;
   }
 
   .features-grid {
